feat(refactor): Modularize main page into tab components
Refactors the monolithic `app/page.tsx` into smaller, feature-specific components (`NewEntryTab`, `HistoryTab`, `GalleryTab`) to improve maintainability, readability, and separation of concerns.

- Extracts state, logic, and JSX for the "New Entry" form into `components/new-entry/NewEntryTab.tsx`.
- Extracts state, data-fetching logic, and JSX for the "History" tab into `components/history/HistoryTab.tsx`.
- Creates a wrapper component `components/gallery/GalleryTab.tsx` for the `MediaGallery`.
- Simplifies `app/page.tsx` to a lightweight container that renders the tab components.

This change fulfills the technical debt refactoring task outlined in the project's TODO list, resulting in a more modular and maintainable codebase without altering user-facing functionality.
