+++ components/gallery/GalleryTab.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import { MediaGallery } from "@/components/MediaGallery"
+
+export function GalleryTab() {
+  return <MediaGallery />
+}
